requirements.txt
torch
open-clip-torch
Pillow
